+++ packages/terra-search-field/src/SearchField.jsx
@@ -220,10 +220,11 @@ class SearchField extends React.Component {
       customProps.className,
     ]);
 
-    const inputText = intl.formatMessage({ id: 'Terra.searchField.search' });
+    const inputText = inputAttributes && Object.prototype.hasOwnProperty.call(inputAttributes, 'aria-label') ? inputAttributes['aria-label'] : intl.formatMessage({ id: 'Terra.searchField.search' });
+
     const buttonText = intl.formatMessage({ id: 'Terra.searchField.submit-search' });
     const clearText = intl.formatMessage({ id: 'Terra.searchField.clear' });
-    const additionalInputAttributes = { 'aria-label': inputText, ...inputAttributes };
+    const additionalInputAttributes = { ...inputAttributes };
     const clearIcon = <span className={cx('clear-icon')} />;
 
     if (value !== undefined) {
@@ -254,6 +255,7 @@ class SearchField extends React.Component {
             placeholder={placeholder}
             onChange={this.handleTextChange}
             disabled={isDisabled}
+            ariaLabel={inputText}
             aria-disabled={isDisabled}
             onKeyDown={this.handleKeyDown}
             refCallback={this.setInputRef}