+++ packages/terra-form-input/CHANGELOG.md
@@ -4,12 +4,20 @@ ChangeLog
 Unreleased
 ----------
 
+2.22.0 - (August 14, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.21.0 - (July 30, 2019)
 ------------------
 ### Changed
 * Minor dependency version bump
 * Updated errorIcon default prop from `Field.defaultProps.errorIcon` with `<IconError>` component.
 * updated package.json test scripts
+
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
 
 2.20.0 - (July 24, 2019)
 ------------------
@@ -18,12 +26,12 @@ Unreleased
 
 2.19.0 - (July 17, 2019)
 ------------------
-## Changed
+### Changed
 * Revert replaced `Field.defaultProps.errorIcon` with `<IconError>` component.
 
 2.18.0 - (July 17, 2019)
 ------------------
-## Changed
+### Changed
 * Replaced `Field.defaultProps.errorIcon` with `<IconError>` component.
 
 2.17.0 - (July 16, 2019)